Move emscripten_console_log and other common library functions
These are all available on all platforms, not just HTML5.

Were trival also move the tests for these function out of
test_browser.py and into test_other.py.
