+++ tests/test_emscripten_unwind_to_js_event_loop.c
@@ -9,7 +9,7 @@ void timeout(void *userData) {
 }
 
 int main() {
-  emscripten_set_timeout(timeout, 2000, 0);
+  emscripten_set_timeout(timeout, 50, 0);
   emscripten_unwind_to_js_event_loop();
   // emscripten_unwind_to_js_event_loop should never return
   __builtin_trap();
